# src/main/resources/static/swagger.yaml
openapi: 3.0.3
info:
  title: Myth Cloth API
  version: 1.0.0
  description: REST API documentation generated from the current controllers and DTOs.
servers:
  - url: /api/v1
    description: Relative to current host
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /figurines:
    get:
      tags: [Figurines]
      summary: List figurines (paginated)
      operationId: retrieveFigurines
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Paginated figurine result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        default:
          $ref: '#/components/responses/ProblemResponse'
    post:
      tags: [Figurines]
      summary: Create figurine
      operationId: createFigurine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FigurineReq'
      responses:
        '201':
          description: Figurine created
          headers:
            Location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FigurineResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
  /figurines/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    get:
      tags: [Figurines]
      summary: Get figurine by id
      operationId: retrieveFigurine
      responses:
        '200':
          description: Figurine
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FigurineResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
    put:
      tags: [Figurines]
      summary: Update figurine
      operationId: updateFigurine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FigurineReq'
      responses:
        '200':
          description: Updated figurine
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FigurineResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
    delete:
      tags: [Figurines]
      summary: Delete figurine
      operationId: deleteFigurine
      responses:
        '204':
          description: Figurine deleted
        default:
          $ref: '#/components/responses/ProblemResponse'
  /figurines/{figurineId}/events:
    parameters:
      - $ref: '#/components/parameters/FigurineIdPathParam'
    get:
      tags: [Figurine Events]
      summary: List figurine events
      operationId: retrieveEvents
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FigurineEventResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
    post:
      tags: [Figurine Events]
      summary: Create figurine event
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FigurineEventReq'
      responses:
        '201':
          description: Event created
          headers:
            Location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FigurineEventResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
  /figurines/{figurineId}/events/{id}:
    parameters:
      - $ref: '#/components/parameters/FigurineIdPathParam'
      - $ref: '#/components/parameters/IdPathParam'
    get:
      tags: [Figurine Events]
      summary: Get figurine event by id
      operationId: retrieveEvent
      responses:
        '200':
          description: Figurine event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FigurineEventResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
    put:
      tags: [Figurine Events]
      summary: Update figurine event
      operationId: updateEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FigurineEventReq'
      responses:
        '200':
          description: Updated figurine event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FigurineEventResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
    delete:
      tags: [Figurine Events]
      summary: Delete figurine event
      operationId: removeEvent
      responses:
        '204':
          description: Event deleted
        default:
          $ref: '#/components/responses/ProblemResponse'
...
  /roles:
    get:
      tags: [Roles]
      summary: List roles
      operationId: retrieveRoles
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
    post:
      tags: [Roles]
      summary: Create role
      operationId: createRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleReq'
      responses:
        '201':
          description: Role created
          headers:
            Location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
  /roles/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    get:
      tags: [Roles]
      summary: Get role by id
      operationId: retrieveRole
      responses:
        '200':
          description: Role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
    put:
      tags: [Roles]
      summary: Update role
      operationId: updateRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleReq'
      responses:
        '200':
          description: Updated role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
...
  /collections/{collectionId}/figurines/{figurineId}:
    parameters:
      - $ref: '#/components/parameters/CollectionIdPathParam'
      - $ref: '#/components/parameters/FigurineIdPathParam'
    put:
      tags: [Collections]
      summary: Add figurine to collection
      operationId: addFigurineToCollection
      responses:
        '200':
          description: Figurine added to collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FigurineResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
components:
  parameters:
    IdPathParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    FigurineIdPathParam:
      in: path
      name: figurineId
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    CollectionIdPathParam:
      in: path
      name: collectionId
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    ProblemResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        timestamp:
          type: string
          format: date-time
        errors:
          type: object
          additionalProperties:
            type: string
        error:
          type: string
...
    AnniversaryResp:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        year:
          type: integer
...
    CatalogResp:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
...
    DistributorResp:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        countryCode:
          type: string
          enum: [JP, MX, ES, US, CN]
        website:
          type: string
    FigurineDistributorReq:
      type: object
      properties:
        supplierId:
          type: integer
          format: int64
          minimum: 1
        currency:
          type: string
          enum: [JPY, MXN, EUR, USD, CNY, CAD]
        price:
          type: number
          format: double
          minimum: 0.01
        announcedAt:
          type: string
          format: date
        preorderOpensAt:
          type: string
          format: date
        releaseDate:
          type: string
          format: date
        releaseDateConfirmed:
          type: boolean
          default: false
    FigurineReq:
      type: object
      required: [name, distributors, lineUpId, seriesId, groupId]
      properties:
        name:
          type: string
          maxLength: 100
        distributors:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/FigurineDistributorReq'
        tamashiiUrl:
          type: string
        distributionId:
          type: integer
          format: int64
        lineUpId:
          type: integer
          format: int64
        seriesId:
          type: integer
          format: int64
        groupId:
          type: integer
          format: int64
        anniversaryId:
          type: integer
          format: int64
        isMetalBody:
          type: boolean
        isOriginalColorEdition:
          type: boolean
        isRevival:
          type: boolean
        isPlainCloth:
          type: boolean
        isBattleDamaged:
          type: boolean
        isGoldenArmor:
          type: boolean
        isGold24kEdition:
          type: boolean
        isMangaVersion:
          type: boolean
        isMultiPack:
          type: boolean
        isArticulable:
          type: boolean
        notes:
          type: string
        officialImageUrls:
          type: array
          items:
            type: string
        unofficialImageUrls:
          type: array
          items:
            type: string
    FigurineDistributorResp:
      type: object
      properties:
        distributor:
          $ref: '#/components/schemas/DistributorResp'
        currency:
          type: string
          enum: [JPY, MXN, EUR, USD, CNY, CAD]
        price:
          type: number
          format: double
        priceWithTax:
          type: number
          format: double
        announcedAt:
          type: string
          format: date
        preorderOpensAt:
          type: string
          format: date
        releaseDate:
          type: string
          format: date
        releaseDateConfirmed:
          type: boolean
    FigurineEventReq:
      type: object
      required: [description, date, region, type, figurineId]
      properties:
        description:
          type: string
          maxLength: 100
        date:
          type: string
          format: date
        region:
          type: string
          enum: [JP, MX, ES, US, CN]
        type:
          type: string
          enum: [ANNOUNCEMENT, PREORDER_OPEN, PREORDER_CLOSE, RELEASE, RESTOCK, LOCAL_CONFIRMATION, LOCAL_RELEASE]
        figurineId:
          type: integer
          format: int64
          minimum: 1
    FigurineEventResp:
      type: object
      properties:
        id:
          type: integer
          format: int64
        date:
          type: string
          format: date
        type:
          type: string
          enum: [ANNOUNCEMENT, PREORDER_OPEN, PREORDER_CLOSE, RELEASE, RESTOCK, LOCAL_CONFIRMATION, LOCAL_RELEASE]
        region:
          type: string
          enum: [JP, MX, ES, US, CN]
        description:
          type: string
...
    FigurineResp:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        displayableName:
          type: string
        distributors:
          type: array
          items:
            $ref: '#/components/schemas/FigurineDistributorResp'
        tamashiiUrl:
          type: string
        distribution:
          $ref: '#/components/schemas/CatalogResp'
        lineUp:
          $ref: '#/components/schemas/CatalogResp'
        series:
          $ref: '#/components/schemas/CatalogResp'
        group:
          $ref: '#/components/schemas/CatalogResp'
        anniversary:
          $ref: '#/components/schemas/AnniversaryResp'
        isMetalBody:
          type: boolean
        isOriginalColorEdition:
          type: boolean
        isRevival:
          type: boolean
        isPlainCloth:
          type: boolean
        isBattleDamaged:
          type: boolean
        isGoldenArmor:
          type: boolean
        isGold24kEdition:
          type: boolean
        isMangaVersion:
          type: boolean
        isMultiPack:
          type: boolean
        isArticulable:
          type: boolean
        notes:
          type: string
        officialImageUrls:
          type: array
          items:
            type: string
        unofficialImageUrls:
          type: array
          items:
            type: string
        events:
          type: array
          items:
            $ref: '#/components/schemas/FigurineEventResp'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaginatedResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/FigurineResp'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...
    RoleReq:
      type: object
      required: [description]
      properties:
        description:
          type: string
          maxLength: 200
    RoleResp:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
...